requirements.txt
requests
appdirs
aria2p
progress
psutil
bs4
pytz
Pillow>=4.2.0
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
js2py
python-dotenv
tenacity
python-magic
beautifulsoup4
TgCrypto
youtube_dl
megasdkrestclient>=0.1.1,<1.0.0
feedparser
natsort
heroku3
aiohttp
psycopg2-binary
git+https://github.com/yt-dlp/yt-dlp
lxml
telegraph
speedtest-cli
messages
oauth2client
progressbar2
httplib2shim
google_auth_oauthlib
pyrogram
